pyproject.toml
[tool.poetry]
name = "webhook-receiver"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
redis = "^5.2.1"
uvicorn = "^0.34.0"
celery = "^5.4.0"
pyyaml = "^6.0.2"
httpx = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
